release version 1.0

+++ lib/database/queries/reminder/reminder.dart
@@ -1,5 +1,7 @@
 import 'package:hive/hive.dart';
 import 'package:remind_me_of/database/models/reminder/reminder.dart';
+import 'package:remind_me_of/database/models/settings/settings.dart';
+import 'package:remind_me_of/database/queries/settings/settings.dart';
 import 'package:remind_me_of/services/local_notification.dart';
 
 Future<void> createReminder({required Reminder reminder}) async {
@@ -10,6 +12,7 @@ Future<void> createReminder({required Reminder reminder}) async {
 
 Future<List<Reminder>> getReminderList({String ?category}) async {
   var box = await Hive.openBox<Reminder>('reminder');
+  Settings settings = await getSettings();
   List<Reminder> reminderList = [];
   DateTime now = DateTime.now();
 
@@ -17,7 +20,7 @@ Future<List<Reminder>> getReminderList({String ?category}) async {
     var list = box.values.where((reminder) => reminder.category == category);
 
     list.forEach((value) async {
-      if(value.reminderDate.isBefore(now)){
+      if(value.reminderDate.isBefore(now) && settings.deletePastReminders){
         if(value.repeat == '' || value.repeat == 'not_repeat'){
           await LocalNotificationService.cancelNotification(id: value.key);
           await deleteReminder(key: value.key);
@@ -29,7 +32,7 @@ Future<List<Reminder>> getReminderList({String ?category}) async {
 
   }else{
     box.toMap().forEach((key, value) async {
-      if(value.reminderDate.isBefore(now)){
+      if(value.reminderDate.isBefore(now) && settings.deletePastReminders){
         if(value.repeat == '' || value.repeat == 'not_repeat'){
           await LocalNotificationService.cancelNotification(id: value.key);
           await deleteReminder(key: value.key);
